# Remove runtime dependency on setuptools