--- a/setup.cfg
+++ b/setup.cfg
@@ -50,7 +50,6 @@ install_requires =
    uncertainties>=3.1.4
    h5netcdf>=0.10.0,!=0.14.0
 ; see   https://github.com/h5netcdf/h5netcdf/issues/154
-   setuptools>=48
    versioningit>=1.1.0
 ;  transitive dependencies. We list these explicitly to
 ;  ensure that we always use versions that do not have
@@ -58,7 +57,6 @@ install_requires =
    ipython>=7.31.1,!=8.0.0
    pillow>=9.0.0
    rsa>=4.7
-
 
 [options.package_data]
 qcodes =
@@ -88,7 +86,6 @@ test =
     pytest-rerunfailures>=5.0.0
     lxml>=4.3.0
     types_requests>=0.1.8
-    types-setuptools>=57.0.0
     types-tabulate>=0.1.0
     Sphinx>=4.1.2