--- a/setup.cfg
+++ b/setup.cfg
@@ -50,7 +50,7 @@ install_requires =
    uncertainties>=3.1.4
    h5netcdf>=0.10.0,!=0.14.0
 ; see   https://github.com/h5netcdf/h5netcdf/issues/154
-   versioningit>=1.1.0
+   versioningit>=1.1.1
 ;  transitive dependencies. We list these explicitly to
 ;  ensure that we always use versions that do not have
 ;  known security vulnerabilities