--- a/test_requirements.txt
+++ b/test_requirements.txt
@@ -46,7 +46,6 @@ types-ipaddress
 types-pyOpenSSL
 types-pytz
 types-requests
-types-setuptools
 types-six
 types-tabulate
 types-toml