@jenshnielsen 2022-04-19
> do the benchmarks in the description tell us that by removing setuptools as dependency, which is a good thing on it's own, we did not get a slowdown for sure, but also did not get any speed-ups? am i reading it right?

Yes with the fixed version of versioningit (1.1.1) we are not getting the slowdown. Its not a surprise that it performs roughly the same since both versions must inspect the same set of packages. The only difference is that importlib-metadata does this at runtime but setuptools (or rather pgk_resources) does it at import time.